chore(account): render token image in center of breakdown chart
