+++ src/renderer/account/components/AccountPanel/AccountPanel.scss
@@ -13,6 +13,7 @@
     .breakdown {
       flex: 1 1 auto;
       min-height: 400px;
+      height: 100%;
     }
 
     .address {
