+++ src/renderer/account/components/AccountPanel/Breakdown/BreakdownChart/BreakdownChart.scss
@@ -1,0 +1,12 @@
+.breakdownChart {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+  position: relative;
+
+  text {
+    fill: #494759;
+    font-family: $primary-font-stack;
+  }
+}
